refactor(backend): move method from media to misc service
